move tokenize to dedicated submodule
