+++ dask/array/ma.py
@@ -9,7 +9,7 @@ from dask.array.core import asanyarray, blockwise, elemwise, map_blocks
 from dask.array.reductions import reduction
 from dask.array.routines import _average
 from dask.array.routines import nonzero as _nonzero
-from dask.base import normalize_token
+from dask.tokenize import normalize_token
 from dask.utils import derived_from
