+++ dask/array/ufunc.py
@@ -7,8 +7,9 @@ import numpy as np
 
 from dask import core
 from dask.array.core import Array, apply_infer_dtype, asarray, blockwise, elemwise
-from dask.base import is_dask_collection, normalize_token
+from dask.base import is_dask_collection
 from dask.highlevelgraph import HighLevelGraph
+from dask.tokenize import normalize_token
 from dask.utils import derived_from, funcname
